+++ src/Algorithm/Clustering.cpp
@@ -5,7 +5,7 @@ namespace algorithm
 {
     //For Dynamic Vector
     void KMedoidsClusteringDynamic(const geometry::PointXList & wait_to_cluster, 
-        std::vector<ClusterDynamic> &clustering_result, int target_number, bool initialized, 
+        std::vector<ClusterDynamic> &clustering_result, size_t target_number, bool initialized, 
         const std::vector<int> & initialized_index)
     {
         //randomly pick $target_number points as the start medoids
@@ -19,7 +19,7 @@ namespace algorithm
         }
         else
         {
-            for(int i = 0; i != wait_to_cluster.size(); ++i)
+            for(size_t i = 0; i != wait_to_cluster.size(); ++i)
             indexs.push_back(i);
             std::random_shuffle(indexs.begin(), indexs.end());
             indexs.resize(target_number);
@@ -32,17 +32,17 @@ namespace algorithm
             //std::cout<<"iteration "<<iteration<<std::endl;
             std::vector<bool> no_change(target_number, false);
             iteration += 1;
-            for(int i = 0; i != target_number; ++i)
+            for(size_t i = 0; i != target_number; ++i)
             {
                 cluster_index[i].clear();
                 //cluster_index[i].push_back(indexs[i]);
             }
             //std::cout<<"choose new medoid "<<std::endl;
-            for(int i = 0; i != wait_to_cluster.size(); ++i)
+            for(size_t i = 0; i != wait_to_cluster.size(); ++i)
             {
                 int min_index = -1;
                 float min_distance = std::numeric_limits<float>::max ();
-                for(int j = 0; j != target_number; ++j)
+                for(size_t j = 0; j != target_number; ++j)
                 {
                     float distance = (wait_to_cluster[i] - wait_to_cluster[indexs[j]]).norm();
                     if( distance < min_distance)
@@ -56,15 +56,15 @@ namespace algorithm
             }
             //choose new medoid
 
-            for(int i = 0; i != target_number; ++i)
+            for(size_t i = 0; i != target_number; ++i)
             {
                 float min_distance_sum = std::numeric_limits<float>::max ();
                 int new_medoid_index = -1;
                 auto &current_cluster_index = cluster_index[i];
-                for(int test_id = 0; test_id != current_cluster_index.size(); ++test_id)
+                for(size_t test_id = 0; test_id != current_cluster_index.size(); ++test_id)
                 {
                     float distance_sum = 0;
-                    for(int j = 0; j != current_cluster_index.size(); ++j)
+                    for(size_t j = 0; j != current_cluster_index.size(); ++j)
                     {
                         distance_sum +=
                             (wait_to_cluster[current_cluster_index[j]] -wait_to_cluster[current_cluster_index[test_id]]).norm();
@@ -80,7 +80,7 @@ namespace algorithm
                 else indexs[i] = new_medoid_index;
             }
             bool all_no_change = true;
-            for(int i = 0; i != target_number; ++i)
+            for(size_t i = 0; i != target_number; ++i)
             {
                 all_no_change = all_no_change & no_change[i];
             }
@@ -88,12 +88,12 @@ namespace algorithm
             if(all_no_change == true)
             break;
         }
-        for(int i = 0; i != target_number; ++i)
+        for(size_t i = 0; i != target_number; ++i)
         {
              
             clustering_result.push_back(ClusterDynamic(wait_to_cluster[indexs[i]]));
             clustering_result.back().indexs = cluster_index[i];
-            for(int j = 0; j != cluster_index[i].size(); ++j)
+            for(size_t j = 0; j != cluster_index[i].size(); ++j)
                 clustering_result.back().items.push_back(wait_to_cluster[cluster_index[i][j]]);
         }
     }
